fastapi/requirements.txt
fastapi
uvicorn
pymongo
kafka-python
python-dotenv
